CDR-2022 move TestDataTemplateProvider
