+++ serialisation/src/test/java/org/ehrbase/openehr/sdk/serialisation/TestDataTemplateProvider.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.ehrbase.openehr.sdk.serialisation.flatencoding;
+package org.ehrbase.openehr.sdk.serialisation;
 
 import java.io.IOException;
 import java.util.Arrays;
