+++ serialisation/src/test/java/org/ehrbase/openehr/sdk/serialisation/dto/GeneratedDtoToRmConverterTest.java
@@ -61,7 +61,7 @@ import org.ehrbase.openehr.sdk.generator.commons.test_data.dto.korpergrossecompo
 import org.ehrbase.openehr.sdk.generator.commons.test_data.dto.korpergrossecomposition.definition.GrosseLangeObservation;
 import org.ehrbase.openehr.sdk.generator.commons.test_data.dto.testalltypesenv1composition.TestAllTypesEnV1Composition;
 import org.ehrbase.openehr.sdk.serialisation.RMDataFormat;
-import org.ehrbase.openehr.sdk.serialisation.flatencoding.TestDataTemplateProvider;
+import org.ehrbase.openehr.sdk.serialisation.TestDataTemplateProvider;
 import org.ehrbase.openehr.sdk.test_data.composition.CompositionTestDataCanonicalJson;
 import org.junit.jupiter.api.Test;
